+++ addForm.js
@@ -1,15 +1,15 @@
 
-export const renderForm = () => {
-    const appElement = document.querySelector('#app');
-    const formHtml =`
-    <div class="add-form">
-    <input id="input" type="text" class="add-form-name" placeholder="Введите ваше имя" />
-    <textarea id="textarea" type="textarea" class="add-form-text" placeholder="Введите ваш коментарий"
-    rows="4"></textarea>
-    <div class="add-form-row">
-    <button id="button" class="add-form-button">Написать</button>
-    </div>
-    </div>`;
+// export const renderForm = () => {
+//     const appElement = document.querySelector('#app');
+//     const formHtml =`
+//     <div class="add-form">
+//     <input id="input" type="text" class="add-form-name" placeholder="Введите ваше имя" />
+//     <textarea id="textarea" type="textarea" class="add-form-text" placeholder="Введите ваш коментарий"
+//     rows="4"></textarea>
+//     <div class="add-form-row">
+//     <button id="button" class="add-form-button">Написать</button>
+//     </div>
+//     </div>`;
 
-    appElement.innerHTML = formHtml;
-}
+//     appElement.innerHTML = formHtml;
+// }
