+++ index.html
@@ -10,11 +10,10 @@
 <body>
   <div class="container">
     <span class="loader-comments hidden" > Пожалуйста подождите, идет загрузка комментариев...</span>
-    <ul id="list" class="comments">
-    </ul>
-    <span class="title-link hidden" id="title-link">Что бы добавить коментарий,
+    <!-- <ul id="list" class="comments"></ul> -->
+    <!-- <span class="title-link hidden" id="title-link">Что бы добавить коментарий,
       <a class="link-authorization" href="#" id="link-authorization">авторизутесь</a>
-    </span>
+    </span> -->
     <span class="loader-newcomment hidden">Загрузка комментария...</span>
     <!-- <div class="add-form">
       <input id="input" type="text" class="add-form-name" placeholder="Введите ваше имя" />
@@ -26,6 +25,7 @@
     </div> -->
     <div id="app"></div>
   </div>
+  
 </body>
 <style>
   .error {
