+++ main.js
@@ -1,43 +1,34 @@
 "use strict";
 
-import { renderForm } from "./addForm.js";
 import { postTodo } from "./api.js";
 import { fetchComments } from "./fetch.js";
 import { changementLikes } from "./likes.js";
-import { renderLogin } from "./loginPage.js";
 import { renderComments } from "./renderСomments.js";
 
-  const buttonEl = document.getElementById('button');
-  const inputName = document.querySelector('input');
-  const textareaComment = document.getElementById('textarea');
-  const loaderComments = document.querySelector('.loader-comments');
-  const loaderNewcomment = document.querySelector('.loader-newcomment');
-  const formEl = document.querySelector('.add-form');
-  const linkAuthotization = document.querySelector('.link-authorization');
-  const listCart = document.getElementById('list');
-  const titleLink = document.querySelector('.title-link');
+const loaderComments = document.querySelector('.loader-comments');
+
 
   loaderComments.classList.remove("hidden")
 
   fetchComments();
 
-  linkAuthotization.addEventListener('click', () => {
-    listCart.classList.add("hidden");
-    titleLink.classList.add("hidden");
-    renderLogin();
-  });
-
   let comments = [];
 
-  renderComments({comments});
+  // renderComments({comments});
 
-    function protectInput(text) {
+    export function protectInput(text) {
       return text.replaceAll('<', '&lt').replaceAll('>', '&gt')
       .replaceAll('QUOTE_BEGIN', '<div class="quote">').replaceAll('QUOTE_END', '</div>');
     }
 
-buttonEl.setAttribute('disabled', true);
-inputName.addEventListener('input', function (e) {
+    const loaderNewcomment = document.querySelector('.loader-newcomment');
+
+    const buttonEl = document.getElementById('button');
+    const inputName = document.querySelector('#input');
+    const textareaComment = document.getElementById('textarea');
+
+  buttonEl.setAttribute('disabled', true);
+  inputName.addEventListener('input', function (e) {
   if (inputName.value.trim() === '') {
     buttonEl.setAttribute('disabled', true);
   } else {
@@ -46,7 +37,6 @@ inputName.addEventListener('input', function (e) {
 });
 
   buttonEl.addEventListener('click', () => {
-    
     inputName.classList.remove("error");
     textareaComment.classList.remove("error");
     if (inputName.value === '' || textareaComment.value === '') {
