+++ renderСomments.js
@@ -1,11 +1,15 @@
 import { getDate } from "./getDate.js";
 import { changementLikes } from "./likes.js";
 import { answerComment } from "./answerComment.js";
+import { user, token, postTodo } from "./api.js";
+import { renderLogin } from "./loginPage.js";
+import { protectInput } from "./main.js";
+import { fetchComments } from "./fetch.js";
 
-const listCart = document.querySelector('#list');
+
 
 export const renderComments = ({comments}) => {
-  // listCart.classList.remove("hidden");
+  const appElement = document.querySelector('#app');
     const commentsHtml = comments.map((comment, index) => {
       return `<li class="comment">
         <div class="comment-header">
@@ -26,7 +30,43 @@ export const renderComments = ({comments}) => {
       </li>`;
     })
     .join('');
-    listCart.innerHTML = commentsHtml;
-    changementLikes({comments}, {renderComments});
-    answerComment({comments});
+
+    const appHtml = `
+    <ul id="list" class="comments">${commentsHtml}</ul>
+    <div class="add-form">
+      <input id="input" type="text" readonly class="add-form-name" placeholder=${user} />
+      <textarea id="textarea" type="textarea" class="add-form-text" placeholder="Введите ваш коментарий"
+        rows="4"></textarea>
+      <div class="add-form-row">
+        <button id="button" class="add-form-button">Написать</button>
+      </div>
+    </div>`
+  
+    const appNoToken = `
+    <ul id="list" class="comments">${commentsHtml}</ul>
+    <span class="title-link" id="title-link">Что бы добавить коментарий,&nbsp;
+      <a class="link-authorization" href="#" id="link-authorization">авторизутесь</a>
+    </span>`
+
+    if(token) {
+      appElement.innerHTML = appHtml;
+      changementLikes({comments}, {renderComments});
+      answerComment({comments});
+      
+    } else {
+      appElement.innerHTML = appNoToken;
+    }
+
+    // inputName.setAttribute('disabled', 'disabled')
+
+    const linkAuthotization = document.querySelector('.link-authorization');
+    linkAuthotization.addEventListener('click', () => {
+      const listCart = document.getElementById('list');
+      const titleLink = document.querySelector('.title-link');
+      listCart.classList.add("hidden");
+      titleLink.classList.add("hidden");
+      renderLogin();
+    });
+
+
   };
